refactor(ai): extract reusable Insights/Journal review bodies
Pull the Insights AI Review and AI Journal Review renderers out of their
panels into standalone InsightsReviewBody and JournalReviewBody components,
and render them in the AI report history modal for the two new kinds.
